security

+++ src/main/java/mvc/spring/web/controllers/ClubController.java
@@ -14,12 +14,16 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.bind.support.SessionStatus;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import jakarta.validation.Valid;
 import javassist.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import mvc.spring.web.dto.ClubDto;
+import mvc.spring.web.models.UserEntity;
+import mvc.spring.web.security.SecurityUtil;
 import mvc.spring.web.services.club.impl.ClubServiceImpl;
+import mvc.spring.web.services.user.UserService;
 
 @Controller
 @RequestMapping("/clubs")
@@ -28,17 +32,49 @@ import mvc.spring.web.services.club.impl.ClubServiceImpl;
 public class ClubController {
 
     private final ClubServiceImpl clubService;
+    private final UserService userService;
 
     @GetMapping("/all")
     public String listClubs(@NotNull Model model) {
         List<ClubDto> clubs = clubService.findAll();
+        UserEntity user = new UserEntity();
+        String username = SecurityUtil.getSessionUser();
+        if (username != null) {
+            user = userService.findByUsername(username);
+            model.addAttribute("user", user);
+        }
+        model.addAttribute("user", user);
         model.addAttribute("clubs", clubs);
         return "clubs/clubs-list";
+    }
+
+    @GetMapping("/{clubId}")
+    public String clubDetail(@PathVariable("clubId") long clubId, Model model) throws NotFoundException {
+
+        ClubDto club = clubService.findById(clubId);
+        UserEntity user = new UserEntity();
+        String username = SecurityUtil.getSessionUser();
+        if (username != null) {
+            user = userService.findByUsername(username);
+            model.addAttribute("user", user);
+        }
+        model.addAttribute("user", user);
+        model.addAttribute("club", club);
+
+        return "clubs/clubs-detail";
+
     }
 
     @GetMapping("/search")
     public String searchClub(@RequestParam(value = "query") String query, Model model) {
         List<ClubDto> foundClubs = clubService.search(query);
+        UserEntity user = new UserEntity();
+        String username = SecurityUtil.getSessionUser();
+        if (username != null) {
+            user = userService.findByUsername(username);
+            model.addAttribute("user", user);
+        }
+        model.addAttribute("user", user);
         model.addAttribute("clubs", foundClubs);
         return "clubs/clubs-list";
     }
@@ -64,19 +100,35 @@ public class ClubController {
 
     @GetMapping("/{clubId}/edit")
     public String editClubForm(@PathVariable("clubId") long clubId,
-            @NotNull Model model) throws NotFoundException {
-        ClubDto club = clubService.findById(clubId);
-        model.addAttribute("club", club);
-        return "clubs/clubs-edit";
+            RedirectAttributes redirectAttributes,
+            @NotNull Model model) {
+        try {
+            ClubDto club = clubService.findById(clubId);
+            if (!club.getCreatedBy().getUsername().equals(SecurityUtil.getSessionUser())) {
+                redirectAttributes.addFlashAttribute("error", "You don't have permission to update this club.");
+                return "redirect:/clubs/all";
+            }
+            model.addAttribute("club", club);
+            return "clubs/clubs-edit";
+        } catch (NotFoundException e) {
+            redirectAttributes.addFlashAttribute("error", "Club Not Found.");
+            return "redirect:/clubs/all";
+        }
     }
 
     @PostMapping("/{clubId}/edit")
     public String updateClub(@PathVariable("clubId") long clubId,
-            @Valid @ModelAttribute("club") ClubDto club, SessionStatus status,
+            @Valid @ModelAttribute("club") ClubDto club,
+            SessionStatus status,
+            RedirectAttributes redirectAttributes,
             BindingResult result) {
 
         if (result.hasErrors()) {
             return "clubs/clubs-edit";
+        }
+        if (!club.getCreatedBy().getUsername().equals(SecurityUtil.getSessionUser())) {
+            redirectAttributes.addFlashAttribute("error", "You don't have permission to update this club.");
+            return "redirect:/clubs/all";
         }
         club.setId(clubId);
         clubService.update(club);
@@ -84,21 +136,21 @@ public class ClubController {
         return "redirect:/clubs/all";
     }
 
-    @GetMapping("/{clubId}")
-    public String clubDetail(@PathVariable("clubId") long clubId, Model model) throws NotFoundException {
+    @PostMapping("/{clubId}/delete")
+    public String deleteClub(@PathVariable long clubId, RedirectAttributes redirectAttributes) {
+        try {
+            ClubDto club = clubService.findById(clubId);
+            if (!club.getCreatedBy().getUsername().equals(SecurityUtil.getSessionUser())) {
+                redirectAttributes.addFlashAttribute("error", "You don't have permission to delete this club.");
+                return "redirect:/clubs/all";
+            }
+            clubService.deleteById(clubId);
+            return "redirect:/clubs/all";
+        } catch (NotFoundException e) {
 
-        ClubDto club = clubService.findById(clubId);
-        model.addAttribute("club", club);
-
-        return "clubs/clubs-detail";
-
-    }
-
-    @GetMapping("/{clubId}/delete")
-    public String deleteClub(@PathVariable long clubId) {
-        clubService.deleteById(clubId);
-        return "redirect:/clubs/all";
-
+            redirectAttributes.addFlashAttribute("error", "Club not found.");
+            return "redirect:/clubs/all";
+        }
     }
 
 }
